fix: 修复 PlayGameState 中 Go 文字淡出的整数除法 bug，移除 GameSystem 内的冗余自引用

+++ src/main/java/com/likanug/dual/state/PlayGameState.java
@@ -41,7 +41,7 @@ public class PlayGameState extends GameSystemState {
     public void displayMessage(GameSystem system) {
         int messageDurationFrameCount = FPS;
         if (properFrameCount >= messageDurationFrameCount) return;
-        app.fill(0, (float) (255.0 * (1.0 - properFrameCount / messageDurationFrameCount)));
+        app.fill(0, (float) (255.0 * (1.0 - (float) properFrameCount / messageDurationFrameCount)));
         app.text("Go", 0.0F, 0.0F);
     }
